# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SunNXT Clone — API Reference & Security Research",
    "version": "2.0.0",
    "description": "Internal proxy API for the SunNXT clone player, plus documentation of discovered SunNXT upstream API endpoints used for authorized security research.\n\n## Security Findings Summary\n\n| ID | Endpoint | Vulnerability | Severity |\n|---|---|---|---|\n| VULN-11 | `pwaapi.sunnxt.com/licenseproxy/v3/modularLicense/` | No subscription check on DRM license | **Critical** |\n| VULN-12 | `movies1-suntvvod.akamaized.net/.../seg-N.m4s` | Segments served without any auth token | **High** |\n| VULN-13 | Akamai `hdntl` token | Wildcard `acl=/*` covers all content for 24h | **High** |\n| VULN-01 | Login API | Static AES key hardcoded in JS bundle | High |\n\nSee `SECURITY_REPORT.md` for the full report.",
    "contact": {
      "name": "SunNXT Clone — Security Research"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev server" }
  ],
...
  "paths": {
...
    "/api/content/{contentId}": {
      "get": {
        "tags": ["Content"],
        "summary": "Get content metadata",
        "description": "Returns metadata for a piece of content: title, description, images, cast, genres, and subtitle tracks. Does not include stream URLs (see `/api/media/{contentId}`).",
        "parameters": [
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "description": "SunNXT numeric content ID",
            "schema": { "type": "string", "example": "251833" }
          }
        ],
        "responses": {
          "200": {
            "description": "Content metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContentResponse" }
              }
            }
          },
          "500": {
            "description": "Upstream error — returns empty results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "array", "items": {}, "example": [] }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/{contentId}": {
      "get": {
        "tags": ["Playback"],
        "summary": "Get stream URLs for a content ID",
        "description": "Fetches the media API response for a content ID, returning stream URLs (DASH/HLS), license URLs, subtitle tracks, and related metadata. Handles AES-CBC response decryption, geo-block detection, session refresh, and falls back to `pwaapi` if the main endpoint returns a subscription error.",
        "parameters": [
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "description": "SunNXT numeric content ID",
            "schema": { "type": "string", "example": "115249" }
          }
        ],
        "responses": {
          "200": {
            "description": "Media data with stream URLs",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MediaResponse" }
              }
            }
          },
          "401": {
            "description": "Not logged in or session refresh failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 401 },
                    "error": { "type": "string", "example": "login_required" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Content requires a subscription",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 404 },
                    "error": { "type": "string", "example": "video_unavailable" },
                    "message": { "type": "string", "example": "Please subscribe to watch the content" }
                  }
                }
              }
            }
          },
          "451": {
            "description": "Geo-blocked — international roaming expired or IP not India",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 451 },
                    "error": { "type": "string", "example": "geo_blocked" },
                    "title": { "type": "string" },
                    "message": { "type": "string" },
                    "blocked_reason": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid",
        "description": "SunNXT session cookie obtained from POST /api/auth/login"
      }
    },
    "schemas": {
...
      "ContentItem": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "globalServiceName": { "type": "string" },
          "releaseDate": { "type": "string", "format": "date" },
          "generalInfo": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "description": { "type": "string" },
              "type": { "type": "string", "example": "movie" },
              "isDolby": { "type": "boolean" }
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": { "type": "string", "example": "poster" },
                "values": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "resolution": { "type": "string", "example": "xhdpi" },
                      "link": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ContentResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ContentItem" }
          }
        }
      },
      "VideoEntry": {
        "type": "object",
        "properties": {
          "link": {
            "type": "string",
            "description": "Stream URL — DASH manifest (.mpd) or HLS playlist (.m3u8). May contain Akamai `hdnea` auth token.",
            "example": "https://suntvvod1.sunnxt.com/.../251833_hd.mpd?hdnea=..."
          },
          "licenseUrl": {
            "type": "string",
            "description": "Widevine license server URL. Present only for DRM-protected streams.",
            "example": "https://api.sunnxt.com/nagravision/license?content_id=115249"
          },
          "format": {
            "type": "string",
            "description": "Stream format identifier",
            "enum": ["dash", "dash-cenc", "hls", "hlsaes", "hls-fp-aapl"],
            "example": "dash"
          },
          "profile": { "type": "string", "example": "hd" },
          "resolution": { "type": "string", "example": "1280x720" }
        }
      },
      "MediaResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "videos": {
                  "type": "object",
                  "properties": {
                    "values": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/VideoEntry" }
                    }
                  }
                },
                "subtitles": {
                  "type": "object",
                  "properties": {
                    "values": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "language": { "type": "string", "example": "Tamil" },
                          "link": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}